test(core): guard GITATTRIBUTES_EOL_RULES ↔ .gitattributes sync; extract to module
The EOL pins shipped incomplete because two sources drifted apart with no test
to catch it: the injector list in cli.js and n-dx's own .gitattributes. Fixing
today's three missing patterns doesn't stop the next divergence.

- Extract GITATTRIBUTES_EOL_RULES / HEADER / ensureGitattributesRules into
  packages/core/gitattributes-pins.js as a single importable source of truth
  (cli.js self-executes, so a test can't import it directly).
- Add a sync-guard test (prd-line-endings.test.js) asserting the injector's
  eol=lf pattern set EQUALS n-dx's own .gitattributes pattern set. Any future
  divergence fails CI — verified with a negative test (dropping one pattern
  makes the guard red with a readable diff).

No behavior change; refactor + test only.
